Major commit
Multiple input maps, safe checks for input values, bug fixes for weights
update procedure and dropout

+++ c++/sources/layer_f.h
@@ -45,8 +45,8 @@ public:
 private:
   Weights weights_;
   Weights biases_;
-  float droprate_;  
-  float c_; // coefficient for SVM
+  double droprate_;  
+  double c_; // coefficient for SVM
   
 };
